“Confusion Matrix” (using a cut-off value for prediction at 0.5)

based on this cut-off value, create the confusion matrix and also output the following metrics: prevalence, accuracy, sensitivity, specificity, false discovery rate, and diagnostic odds ratio

confusion_matrix(Data, X, y, cut_off = 0.5)

Arguments

y

: A integer denoting the column of Data representing the response variable.

cut_off

: A number of the value for prediction, default is 0.5. (i.e. assign value 1 for predictions above 0.5 and value 0 for prediction below or equal to 0.5)

Data:

A matrix containing the data we want to perform logistic regression on.

X:

A vector denoting the columns of Data representing the predictor variables.

Value

A list of the following attributes: `confusion_matrix`, `prevalence`, `accuracy`, `sensitivity`, `specificity`, `false_discovery_rate`, `diagnostic_odds_ratio`

Author

Group 11

Examples

#data obtained from https://stats.oarc.ucla.edu/r/dae/logit-regression/
test_data <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
  test_resp <- 1
  test_pred <- 2:4
# Call your confusion_matrix function
result <- confusion_matrix(Data = test_data, X=test_pred, y=test_resp, cut_off = 0.5)
print(result)
#> $confuse_matrix
#>                 Predicted Positive Predicted Negative
#> Actual Positive                263                118
#> Actual Negative                 10                  9
#> 
#> $prevalence
#> [1] 0.3175
#> 
#> $accuracy
#> [1] 0.68
#> 
#> $sensitivity
#> [1] 0.07086614
#> 
#> $specificity
#> [1] 0.96337
#> 
#> $false_discovery_rate
#> [1] 0.5263158
#> 
#> $diagnostic_odds_ratio
#> [1] 2.005932
#>