Finding the Betas

This function will spit out the desired beta values for the log loss function

log_betas(Data, Y, X)

Arguments

Data

a matrix of data that we want to do a logistic regression on.

Y

A integer representing the column of the data that you want to be the response variable

X

A vector representing the column(s) of the data that you want to be the predictor variable(s).

Value

A vector giving the optimized beta values

Author

Group 11

Examples

  #data obtained from https://stats.oarc.ucla.edu/r/dae/logit-regression/
  test_data <- read.csv("https://stats.idre.ucla.edu/stat/data/binary.csv")
  test_resp <- 1
  test_pred <- 2:4
  log_betas(Data = test_data, Y = test_resp, X = test_pred)
#> $par
#>              [,1]
#> [1,] -3.440677328
#> [2,]  0.002289743
#> [3,]  0.774910748
#> [4,] -0.559607591
#> 
#> $value
#> [1] 229.7209
#> 
#> $counts
#> function gradient 
#>      473       NA 
#> 
#> $convergence
#> [1] 0
#> 
#> $message
#> NULL
#>